+++ src/main/kotlin/dev/nyon/bbm/KeyBindings.kt
@@ -1,11 +1,6 @@
 package dev.nyon.bbm
 
 import com.mojang.blaze3d.platform.InputConstants
-import dev.nyon.bbm.config.getActiveConfig
-//? if fabric
-import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking
-//? if neoforge
-/*import net.neoforged.neoforge.network.PacketDistributor*/
 import net.minecraft.client.KeyMapping
 import org.lwjgl.glfw.GLFW
 
@@ -14,15 +9,5 @@ object KeyBindings {
         KeyMapping(
             "key.bbm.jump", InputConstants.Type.KEYSYM, GLFW.GLFW_KEY_H, "key.bbm.category"
         )
-    }
-
-    fun handleKeyBindings() {
-        if (jumpKeyBind.isDown && getActiveConfig()?.allowJumpKeybind == true) {
-            /*? if fabric {*/
-            ClientPlayNetworking.send(PressJumpKeybindingPacket)
-            /*?} else {*/
-            /*PacketDistributor.sendToServer(PressJumpKeybindingPacket)
-            *//*?}*/
-        }
     }
 }